backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-telegram-bot==21.6
openai==1.54.0
anthropic==0.39.0
psycopg[binary]==3.2.3
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
jinja2==3.1.4
python-docx==1.1.2
